Refactor e2e fixture structure
