+++ src/test/cypress/e2e/course/CourseExercise.cy.ts
@@ -1,6 +1,6 @@
 import { artemis } from '../../support/ArtemisTesting';
 import { Course } from '../../../../main/webapp/app/entities/course.model';
-import multipleChoiceQuizTemplate from '../../fixtures/quiz_exercise_fixtures/multipleChoiceQuiz_template.json';
+import multipleChoiceQuizTemplate from '../../fixtures/exercise/quiz/multiple_choice/template.json';
 import { convertCourseAfterMultiPart } from '../../support/requests/CourseManagementRequests';
 import { generateUUID } from '../../support/utils';
