+++ src/test/cypress/fixtures/exercise/modeling/template.json
