+++ src/test/cypress/fixtures/exercise/programming/all_successful/Policy.txt
