+++ src/test/cypress/fixtures/exercise/programming/all_successful/submission.json
@@ -1,0 +1,29 @@
+{
+    "files": [
+        {
+            "name": "BubbleSort.java",
+            "path": "exercise/programming/all_successful/BubbleSort.txt"
+        },
+        {
+            "name": "Context.java",
+            "path": "exercise/programming/all_successful/Context.txt"
+        },
+        {
+            "name": "Client.java",
+            "path": "exercise/programming/all_successful/Client.txt"
+        },
+        {
+            "name": "MergeSort.java",
+            "path": "exercise/programming/all_successful/MergeSort.txt"
+        },
+        {
+            "name": "Policy.java",
+            "path": "exercise/programming/all_successful/Policy.txt"
+        },
+        {
+            "name": "SortStrategy.java",
+            "path": "exercise/programming/all_successful/SortStrategy.txt"
+        }
+    ],
+    "expectedResult": "100%"
+}
