+++ src/test/cypress/fixtures/exercise/programming/assessment/submission.json
