+++ src/test/cypress/fixtures/exercise/programming/build_error/submission.json
@@ -2,7 +2,7 @@
     "files": [
         {
             "name": "BubbleSort.java",
-            "path": "programming_exercise_submissions/build_error/BubbleSort.txt"
+            "path": "exercise/programming/build_error/BubbleSort.txt"
         }
     ],
     "expectedResult": "0%"
