+++ src/test/cypress/fixtures/exercise/programming/static_code_analysis/BubbleSort.txt
