+++ src/test/cypress/fixtures/exercise/programming/template.json
