+++ src/test/cypress/fixtures/exercise/quiz/drag_and_drop/background.jpg
