+++ src/test/cypress/fixtures/exercise/quiz/drag_and_drop/question.txt
