+++ src/test/cypress/fixtures/exercise/quiz/multiple_choice/question.txt
