+++ src/test/cypress/fixtures/exercise/quiz/multiple_choice/submission.json
