+++ src/test/cypress/fixtures/exercise/quiz/multiple_choice/template.json
