+++ src/test/cypress/fixtures/exercise/quiz/short_answer/question.txt
