+++ src/test/cypress/fixtures/exercise/quiz/short_answer/submission.json
