+++ src/test/cypress/fixtures/exercise/quiz/short_answer/template.json
