+++ src/test/cypress/fixtures/exercise/quiz/template.json
