+++ src/test/cypress/fixtures/lecture/template.json
