--- src/test/cypress/fixtures/programming_exercise_submissions/all_successful/submission.json
@@ -1,29 +1,0 @@
-{
-    "files": [
-        {
-            "name": "BubbleSort.java",
-            "path": "programming_exercise_submissions/all_successful/BubbleSort.txt"
-        },
-        {
-            "name": "Context.java",
-            "path": "programming_exercise_submissions/all_successful/Context.txt"
-        },
-        {
-            "name": "Client.java",
-            "path": "programming_exercise_submissions/all_successful/Client.txt"
-        },
-        {
-            "name": "MergeSort.java",
-            "path": "programming_exercise_submissions/all_successful/MergeSort.txt"
-        },
-        {
-            "name": "Policy.java",
-            "path": "programming_exercise_submissions/all_successful/Policy.txt"
-        },
-        {
-            "name": "SortStrategy.java",
-            "path": "programming_exercise_submissions/all_successful/SortStrategy.txt"
-        }
-    ],
-    "expectedResult": "100%"
-}
