--- src/test/cypress/fixtures/programming_exercise_submissions/partially_successful/submission.json
@@ -1,17 +1,0 @@
-{
-    "files": [
-        {
-            "name": "BubbleSort.java",
-            "path": "programming_exercise_submissions/all_successful/BubbleSort.txt"
-        },
-        {
-            "name": "MergeSort.java",
-            "path": "programming_exercise_submissions/all_successful/MergeSort.txt"
-        },
-        {
-            "name": "SortStrategy.java",
-            "path": "programming_exercise_submissions/all_successful/SortStrategy.txt"
-        }
-    ],
-    "expectedResult": "46.2%"
-}
